:computer: Refactors Player.add_to_queue to a new Queue module
