+++ lib/live_dj/organizer/player.ex
@@ -1,7 +1,5 @@
 defmodule LiveDj.Organizer.Player do
 
-  alias LiveDj.Organizer.Video
-
   def get_initial_state do
     %{state: "paused", video_id: "", time: 0}
   end
@@ -32,18 +30,6 @@ defmodule LiveDj.Organizer.Player do
       videoId: player.video_id,
       time: player.time
     }
-  end
-
-  def add_to_queue(queue, video) do
-    case queue do
-      []  -> [video]
-      [v] -> [Video.update(v, %{next: video.video_id}) | [Video.update(video, %{previous: v.video_id})]]
-      [v|vs] ->
-        videos = Enum.drop(vs, -1)
-        last_video = Video.update(List.last(vs), %{next: video.video_id})
-        new_video = Video.update(video, %{previous: last_video.video_id})
-        [v | videos ++ [last_video, new_video]]
-    end
   end
 
   defp should_play(player) do
